requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
backports-asyncio-runner==1.2.0
black==25.9.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastapi==0.118.0
filetype==1.2.0
frozenlist==1.7.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.25.2
google-auth==2.41.1
google-genai==1.41.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.32
markupsafe==3.0.3
marshmallow==3.26.1
multidict==6.6.4
mypy-extensions==1.1.0
narwhals==2.7.0
numpy==2.2.6
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.0
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.11.0
pydeck==0.9.1
pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
pyyaml==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.25.0
